feat: refine kclvm go gen package and add gengo feature.
